# backend/docs/swagger.yaml
definitions:
...
  domain.MultipleOtherServices:
    properties:
      data:
        items:
          $ref: '#/definitions/domain.OtherService'
        type: array
      meta:
        $ref: '#/definitions/domain.Pagination'
    type: object
...
  domain.OtherService:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: integer
      image_url:
        type: string
      languages:
        items:
          $ref: '#/definitions/domain.OtherServiceTranslation'
        type: array
      updated_at:
        type: string
      website_url:
        type: string
    type: object
  domain.OtherServiceTranslation:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      description:
        type: string
      id:
        type: integer
      language_code:
        type: string
      name:
        type: string
      service_id:
        type: integer
      tag_line:
        type: string
      updated_at:
        type: string
    type: object
  domain.Pagination:
    properties:
      limit:
        type: integer
      next_page:
        type: integer
      offset:
        type: integer
      page:
        type: integer
      previous_page:
        type: integer
      total:
        type: integer
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Testing Swagger APIs.
  termsOfService: http://swagger.io/terms/
  title: Nigus Bingo Desktop APIs
  version: "1.0"
paths:
...
  /other-services:
    get:
      description: Get all other services with pagination, search, sorting, and language
        filters
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Number of services per page
        in: query
        name: limit
        type: integer
      - description: Search term
        in: query
        name: search
        type: string
      - description: Sort by field
        in: query
        name: sort_by
        type: string
      - description: Sort order (asc/desc)
        in: query
        name: sort_order
        type: string
      - collectionFormat: csv
        description: Language codes (comma-separated or repeated)
        in: query
        items:
          type: string
        name: language_codes
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.MultipleOtherServices'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all other services
      tags:
      - Other Services
    post:
      consumes:
      - application/json
      description: Create a new other service with translations
      parameters:
      - description: Other service payload
        in: body
        name: service
        required: true
        schema:
          $ref: '#/definitions/domain.OtherService'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.OtherService'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - JWT: []
      summary: Create a new other service
      tags:
      - Other Services
  /other-services/{id}:
    delete:
      description: Delete an other service by ID
      parameters:
      - description: Service ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - JWT: []
      summary: Delete an other service
      tags:
      - Other Services
    get:
      description: Get a service by ID, optionally filtering translations by language
        codes
      parameters:
      - description: Service ID
        in: path
        name: id
        required: true
        type: integer
      - collectionFormat: csv
        description: Language codes (comma-separated or repeated)
        in: query
        items:
          type: string
        name: language_codes
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.OtherService'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a service by ID
      tags:
      - Other Services
    put:
      consumes:
      - application/json
      description: Update an existing other service
      parameters:
      - description: Service ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated service payload
        in: body
        name: service
        required: true
        schema:
          $ref: '#/definitions/domain.OtherService'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.OtherService'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - JWT: []
      summary: Update an other service
      tags:
      - Other Services
...
securityDefinitions:
  JWT:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"